Overhaul of the entire PKGBUILDs.
Switched from svn to bzr (launchpad) sources, as they are usually
more up to date.

The packets no longer install to /usr/local, but to /usr, which is the
correct place.

Build type is Release now. This shouldn't matter too much.

Documentation is now make'd when doxygen is installed.

Important: There are currently warnings about "references to $srcdir
in the package", and the source of those is known, but I'm not exactly
sure how to fix it, as I have no idea what the offending files do.